requirements.txt
gspread>=5.12.0
google-auth>=2.23.0
requests>=2.31.0
python-dotenv>=1.0.0
rich>=13.7.0
pyperclip>=1.8.2
